examples/todo-api-django-app/api/serializers.py
# -*- coding: utf-8 -*-

from rest_framework import serializers

from .models import Todo


class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = '__all__'
        read_only_fields = ('id',)
