examples/todo-api-django-app/requirements.txt
Django==1.11.7
gunicorn==19.7.1
djangorestframework==3.7.3
python-dotenv==0.7.1
psycopg2==2.7.3.2
requests==2.18.4
pytest==3.2.3
django-cors-headers==2.1.0
bg-kube==0.0.10
